Add new task

+++ src/com/javarush/test/level10/lesson11/home03/Solution.java
@@ -8,11 +8,11 @@ public class Solution
 {
     public static void main(String[] args)
     {
-        int a = 257;
+        int  a = 257;
         int b = 4;
         int c = 3;
         int e = 2;
-        double d =  a + b/c/e;
+        double d = (byte)a + b/c/e;
 
         System.out.println(d);
     }
